pricingservice/src/main/java/com/expeditors/pricingservice/controller/PricingController.java
package com.expeditors.pricingservice.controller;

import com.expeditors.pricingservice.service.PricingService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/pricing")
public class PricingController {

    private final PricingService pricingService;

    public PricingController(PricingService pricingService) {
        this.pricingService = pricingService;
    }

    @GetMapping
    public ResponseEntity<?> getPrice(){
        return ResponseEntity
                .ok()
                .body(pricingService.getPrice());
    }
}
